adds argument to onclick method
a function handling `onclick` events currently does not have any way to tell which image was clicked.  This commit passes the `span` html tag that is the parent of the currently clicked `img` element.  Then we have 
```
onclick = function(tag) {
    let el = $(tag),
         isChecked = el.hasClass("imgChked"),
         imgElement = el.children('img')[0];
}
```
If the image tag has an `id` then we can retrieve it as `imgElement.id`.